+++ test/utils.js
@@ -30,6 +30,11 @@ function makeConfigStub(opts = {}) {
             testRunEnv: NODEJS_TEST_RUN_ENV,
         },
         sets: {},
+        lastFailed: {
+            only: false,
+            input: "some-path",
+            output: "some-other-path",
+        },
     });
 
     const config = {
@@ -38,6 +43,7 @@ function makeConfigStub(opts = {}) {
         system: opts.system,
         sets: opts.sets,
         configPath: opts.configPath,
+        lastFailed: opts.lastFailed,
     };
 
     opts.browsers.forEach(browserId => {